Move vacuum implementations under miio.integrations.vacuum
